Update Structure
Delete entity to create in diferrent classes
