--- shared/src/commonMain/kotlin/com/example/shared/data/remote/IRemoteServicesDataSource.kt
@@ -1,8 +1,0 @@
-package com.example.shared.data.remote
-
-import com.example.shared.domain.entity.Service
-import kotlinx.coroutines.flow.Flow
-
-interface IRemoteServicesDataSource {
-    suspend fun getServicesByWorker(workerId: String): Flow<List<Service>>
-}
