--- shared/src/commonMain/kotlin/com/example/shared/data/repository/IServiceRepository.kt
@@ -1,8 +1,0 @@
-package com.example.shared.data.repository
-
-import com.example.shared.domain.entity.Service
-import kotlinx.coroutines.flow.Flow
-
-interface IServiceRepository {
-    suspend fun getServicesByWorker(workerId: String): Flow<List<Service>>
-}
