+++ shared/src/commonMain/kotlin/com/example/shared/domain/entity/Address.kt
@@ -1,0 +1,14 @@
+package com.example.shared.domain.entity
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Address(
+    val id: String = "",
+    val alias: String = "",
+    val district: String = "",
+    val latitude: Double = 0.0,
+    val longitude: Double = 0.0,
+    val province: String = "",
+    val reference: String = ""
+)
