+++ shared/src/commonMain/kotlin/com/example/shared/domain/entity/Category.kt
@@ -1,0 +1,9 @@
+package com.example.shared.domain.entity
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Category(
+    val id: String = "",
+    val name: String = ""
+)
