+++ shared/src/commonMain/kotlin/com/example/shared/domain/entity/ProfessionalProfileData.kt
@@ -1,0 +1,18 @@
+package com.example.shared.domain.entity
+
+data class ProfessionalProfileData(
+    val workerId: String = "",
+    val uid: String = "",
+    val name: String = "",
+    val email: String = "",
+    val phone: String = "",
+    val profilePictureLink: String = "",
+    val role: String = "",
+    val stars: Double = 0.0,
+    val status: String = "",
+    val travelTime: Int = 0,
+    val trustScore: Int = 0,
+    val locationProvince: String = "",
+    val categoryNames: List<String> = emptyList(),
+    val services: List<Service> = emptyList()
+)
